refactor: abstract encoder-decoder
